Create a TestSemaphore and Lock the getDistance Function in VL6180X.h Bug fixing in Typeconversion

+++ src/microcontroller/FreeRTOS/Src/servo.c
@@ -26,7 +26,7 @@ uint8_t oldAngle = 0;
 
 // Init method
 void Servo_Init(){
-	__HAL_TIM_SET_COMPARE(&htim2, TIM_CHANNEL_2, pwmValue_0); // 0° anfahren
+	__HAL_TIM_SET_COMPARE(&htim2, TIM_CHANNEL_2, (uint16_t)pwmValue_0); // 0° anfahren
 }
 
 // Setter method for servo angle 0..90°
